Add all the logic for use case query price

+++ src/main/java/com/demo/price/domain/model/Amount.java
@@ -3,13 +3,14 @@ package com.demo.price.domain.model;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.Currency;
 
 @Getter
 @Builder
 public class Amount {
 
-    private Long value;
+    private BigDecimal value;
 
     private Currency currency;
 }
